+++ ostwallet/src/main/java/com/ost/ostwallet/App.java
@@ -12,13 +12,14 @@ package com.ost.ostwallet;
 
 import android.app.Application;
 
-import com.datatheorem.android.trustkit.TrustKit;
-import com.ost.ostwallet.network.PersistentCookieStore;
+import com.ost.walletsdk.OstSdk;
 
 import java.net.CookieHandler;
 import java.net.CookieManager;
 import java.net.CookiePolicy;
 import java.net.CookieStore;
+
+import com.ost.ostwallet.network.PersistentCookieStore;
 
 public class App extends Application {
 
@@ -31,8 +32,11 @@ public class App extends Application {
                 CookiePolicy.ACCEPT_ALL);
         CookieHandler.setDefault(cookieManager);
 
-        TrustKit.initializeWithNetworkSecurityConfiguration(getApplicationContext(), R.xml.network_security_config);
         AppProvider.init(getApplicationContext());
         AppProvider.get().setPersistentCookieStore(persistentCookieStore);
+
+        if (null != AppProvider.get().getCurrentEconomy()) {
+            OstSdk.initialize(getApplicationContext(), AppProvider.get().getCurrentEconomy().getSaasApiEndpoint());
+        }
     }
 }