+++ ostwallet/src/main/java/com/ost/ostwallet/ui/dashboard/DashboardActivity.java
@@ -10,7 +10,7 @@
 
 package com.ost.ostwallet.ui.dashboard;
 
-import android.app.AlertDialog;
+import android.app.Dialog;
 import android.content.DialogInterface;
 import android.content.Intent;
 import android.graphics.Color;
@@ -24,19 +24,39 @@ import android.view.View;
 import android.widget.TextView;
 
 import com.crashlytics.android.Crashlytics;
+import com.ost.walletsdk.OstSdk;
+import com.ost.walletsdk.models.entities.OstDevice;
+import com.ost.walletsdk.models.entities.OstUser;
+import com.ost.walletsdk.workflows.OstContextEntity;
+import com.ost.walletsdk.workflows.OstWorkflowContext;
+import com.ost.walletsdk.workflows.errors.OstError;
+import com.ost.walletsdk.workflows.interfaces.OstPinAcceptInterface;
+import com.ost.walletsdk.workflows.interfaces.OstVerifyDataInterface;
+import com.ost.walletsdk.workflows.interfaces.OstWorkFlowCallback;
+
+import org.json.JSONObject;
+
+import java.util.Objects;
+
 import com.ost.ostwallet.AppProvider;
 import com.ost.ostwallet.R;
 import com.ost.ostwallet.entity.Device;
 import com.ost.ostwallet.entity.LogInUser;
 import com.ost.ostwallet.entity.User;
 import com.ost.ostwallet.network.MappyNetworkClient;
+import com.ost.ostwallet.sdkInteract.SdkInteract;
+import com.ost.ostwallet.sdkInteract.WorkFlowListener;
 import com.ost.ostwallet.ui.BaseActivity;
 import com.ost.ostwallet.ui.auth.OnBoardingActivity;
 import com.ost.ostwallet.ui.managedevices.AuthorizeDeviceOptionsFragment;
 import com.ost.ostwallet.ui.managedevices.DeviceListRecyclerViewAdapter;
 import com.ost.ostwallet.ui.workflow.ChildFragmentStack;
+import com.ost.ostwallet.ui.workflow.VerifyDeviceDataFragment;
+import com.ost.ostwallet.ui.workflow.VerifyTransactionDataFragment;
 import com.ost.ostwallet.ui.workflow.WorkFlowPinFragment;
-import com.ost.ostwallet.ui.workflow.createsession.CreateSessionFragment;
+import com.ost.ostwallet.ui.workflow.WorkFlowVerifyDataFragment;
+import com.ost.ostwallet.ui.workflow.recovery.AbortRecoveryFragment;
+import com.ost.ostwallet.ui.workflow.recovery.InitiateRecoveryFragment;
 import com.ost.ostwallet.ui.workflow.transactions.TransactionFragment;
 import com.ost.ostwallet.ui.workflow.walletdetails.WalletDetailsFragment;
 import com.ost.ostwallet.ui.workflow.walletsetup.WalletSetUpFragment;
@@ -44,27 +64,6 @@ import com.ost.ostwallet.util.CommonUtils;
 import com.ost.ostwallet.util.DialogFactory;
 import com.ost.ostwallet.util.FragmentUtils;
 import com.ost.ostwallet.util.KeyBoard;
-import com.ost.walletsdk.OstSdk;
-import com.ost.walletsdk.models.entities.OstToken;
-import com.ost.walletsdk.models.entities.OstUser;
-import com.ost.walletsdk.network.OstApiError;
-import com.ost.walletsdk.ui.OstWalletUI;
-import com.ost.walletsdk.ui.interfaces.FlowCompleteListener;
-import com.ost.walletsdk.ui.interfaces.FlowInterruptListener;
-import com.ost.walletsdk.ui.interfaces.RequestAcknowledgedListener;
-import com.ost.walletsdk.ui.sdkInteract.SdkInteract;
-import com.ost.walletsdk.workflows.OstContextEntity;
-import com.ost.walletsdk.workflows.OstWorkflowContext;
-import com.ost.walletsdk.workflows.errors.OstError;
-import com.ost.walletsdk.workflows.errors.OstErrors;
-import com.ost.walletsdk.workflows.interfaces.OstPinAcceptInterface;
-import com.ost.walletsdk.workflows.interfaces.OstWorkFlowCallback;
-
-import org.json.JSONObject;
-
-import java.math.BigDecimal;
-import java.util.List;
-import java.util.Objects;
 
 import io.fabric.sdk.android.Fabric;
 
@@ -72,9 +71,10 @@ import static android.content.Intent.FLAG_ACTIVITY_NEW_TASK;
 
 public class DashboardActivity extends BaseActivity implements
         TabLayout.OnTabSelectedListener,
-        FlowCompleteListener,
-        FlowInterruptListener,
-        RequestAcknowledgedListener,
+        SdkInteract.FlowComplete,
+        SdkInteract.FlowInterrupt,
+        SdkInteract.PinCallback,
+        SdkInteract.VerifyDataCallback,
         UserListFragment.OnListFragmentInteractionListener,
         WalletSetUpFragment.OnFragmentInteractionListener,
         SettingsFragment.OnFragmentInteractionListener,
@@ -83,8 +83,7 @@ public class DashboardActivity extends BaseActivity implements
         DeviceListRecyclerViewAdapter.OnDeviceListInteractionListener,
         WalletDetailsFragment.OnWalletDetailsFragmentListener,
         AuthorizeDeviceOptionsFragment.OnAuthorizeDeviceOptionsFragmentListener,
-        WalletFragment.walletFragmentInteraction,
-        CreateSessionFragment.OnFragmentInteractionListener {
+        WalletFragment.walletFragmentInteraction {
 
     private static final String LOG_TAG = "DashboardActivity";
     private ViewPager mViewPager;
@@ -117,6 +116,10 @@ public class DashboardActivity extends BaseActivity implements
         Objects.requireNonNull(mTabLayout.getTabAt(1)).setIcon(R.drawable.wallet_icon);
         Objects.requireNonNull(mTabLayout.getTabAt(2)).setIcon(R.drawable.settings_icon);
 
+        //Set SdkInteract Pin and verify data listeners
+        SdkInteract.getInstance().setPinCallbackListener(this);
+        SdkInteract.getInstance().setVerifyDataCallbackListener(this);
+        SdkInteract.getInstance().setFlowListeners(this);
 
         setUpDevice();
 
@@ -133,17 +136,11 @@ public class DashboardActivity extends BaseActivity implements
     private void checkForActiveUserAndDevice() {
         OstUser ostUser = AppProvider.get().getCurrentUser().getOstUser();
         if (!(ostUser.isActivated() || ostUser.isActivating())) {
-            long expiredAfterInSecs = 30 * 24 * 60 * 60;
-            Integer decimals = Integer.parseInt(OstToken.getById(AppProvider.get().getCurrentUser().getTokenId()).getBtDecimals());
-            String spendingLimit = new BigDecimal("1000").multiply(new BigDecimal(10).pow(decimals)).toString();
-            String workflowId = OstWalletUI.activateUser(DashboardActivity.this,
-                    ostUser.getId(),
-                    expiredAfterInSecs,
-                    spendingLimit,
-                    AppProvider.get().getUserPassphraseCallback());
-            OstWalletUI.subscribe(workflowId, this);
+            FragmentUtils.addFragment(R.id.layout_container,
+                    WalletSetUpFragment.newInstance(),
+                    this);
             mViewPager.setCurrentItem(1);
-        } else if(null != ostUser.getCurrentDevice() && ostUser.getCurrentDevice().canBeAuthorized()) {
+        } else if(ostUser.getCurrentDevice().canBeAuthorized()) {
             handleCrashAnalytics();
             mSettingsFragment.setOpenDeviceAuthorization(true);
             mViewPager.setCurrentItem(2);
@@ -201,31 +198,25 @@ public class DashboardActivity extends BaseActivity implements
     }
 
     @Override
-    public void activateAcknowledged(String workflowId) {
+    public void activateAcknowledged(long workflowId) {
         FragmentUtils.goBack(this);
         handleCrashAnalytics();
     }
 
     @Override
-    public void flowComplete(OstWorkflowContext ostWorkflowContext, OstContextEntity ostContextEntity) {
-        String workflowId = ostWorkflowContext.getWorkflowId();
+    public void flowComplete(long workflowId, OstWorkflowContext ostWorkflowContext, OstContextEntity ostContextEntity) {
         showProgress(false);
         if (OstWorkflowContext.WORKFLOW_TYPE.ACTIVATE_USER
                 .equals(
-                        ostWorkflowContext.getWorkflowType()
+                        ostWorkflowContext.getWorkflow_type()
                 )) {
             showUserActivationToast = true;
             notifyActivate();
             return;
-        } else if (OstWorkflowContext.WORKFLOW_TYPE.UPDATE_BIOMETRIC_PREFERENCE
-                .equals(
-                        ostWorkflowContext.getWorkflowType()
-                )) {
-            mSettingsFragment.reDrawView();
         }
         JSONObject trxWorkflow = null;
         trxWorkflow = transactionWorkflows.optJSONObject(String.format("%s", workflowId));
-        String successMessage = new CommonUtils().formatWorkflowSuccessToast(ostWorkflowContext.getWorkflowType(), trxWorkflow);
+        String successMessage = new CommonUtils().formatWorkflowSuccessToast(ostWorkflowContext.getWorkflow_type(), trxWorkflow);
         if(successMessage != null){
             if(trxWorkflow != null){
                 refreshWalletAfter(500);
@@ -252,41 +243,13 @@ public class DashboardActivity extends BaseActivity implements
     }
 
     @Override
-    public void flowInterrupt(OstWorkflowContext ostWorkflowContext, OstError ostError) {
+    public void flowInterrupt(long workflowId, OstWorkflowContext ostWorkflowContext, OstError ostError) {
         showProgress(false);
-
-        Log.d(LOG_TAG, String.format("Flow Interrupted: WorkFlow Id: %s of Workflow %s", ostWorkflowContext.getWorkflowId() ,ostWorkflowContext.getWorkflowType().toString()));
-
-        //region Sdk Error Handling
-        if (OstErrors.ErrorCode.WORKFLOW_CANCELLED == ostError.getErrorCode()) {
-            //Test-App has cancelled the workflow
-            Log.i(LOG_TAG, "Interrupt Reason: Test-App cancelled the workflow.");
-        } else if (ostError.isApiError()) {
-            OstApiError apiError = (OstApiError) ostError;
-            if (apiError.isApiSignerUnauthorized()) {
-                // The device has been revoked and can not make any more calls to OST Platform.
-                // Apps must logout users at this point.
-                // For purpose of testing the sdk, lets give users an option.
-                deviceUnauthorized(apiError);
-            } else {
-                //Let's log the error
-                logSdkError(ostWorkflowContext, ostError);
-            }
-        } else if (OstErrors.ErrorCode.DEVICE_NOT_SETUP == ostError.getErrorCode()) {
-            // Device needs to be registered or new device keys need to be created.
-            // To perform this operation, Test-App needs to call OstSdk.setupDevice()
-            // If app believe that user is authenticated, they should logout user here.
-            deviceUnauthorized(ostError);
-        } else {
-            //Let's log the error
-            logSdkError(ostWorkflowContext, ostError);
-        }
-
         JSONObject trxWorkflow = null;
         try{
-            trxWorkflow = transactionWorkflows.getJSONObject(String.format("%s", ostWorkflowContext.getWorkflowId()));
+            trxWorkflow = transactionWorkflows.getJSONObject(String.format("%s", workflowId));
         } catch (Exception e){}
-        String failMessage = new CommonUtils().formatWorkflowFailedToast(ostWorkflowContext.getWorkflowType(), ostError, trxWorkflow);
+        String failMessage = new CommonUtils().formatWorkflowFailedToast(ostWorkflowContext.getWorkflow_type(), ostError, trxWorkflow);
         if(failMessage != null){
             showToastMessage(failMessage, false);
         }
@@ -294,7 +257,7 @@ public class DashboardActivity extends BaseActivity implements
 
         if (OstWorkflowContext.WORKFLOW_TYPE.ACTIVATE_USER
                 .equals(
-                        ostWorkflowContext.getWorkflowType()
+                        ostWorkflowContext.getWorkflow_type()
                 )) {
             Log.e(LOG_TAG, "User Activate failed");
             checkForActiveUserAndDevice();
@@ -308,6 +271,23 @@ public class DashboardActivity extends BaseActivity implements
                 this);
     }
 
+    @Override
+    public void getPin(long workflowId, OstWorkflowContext ostWorkflowContext, String userId, OstPinAcceptInterface ostPinAcceptInterface) {
+        showProgress(false);
+        showGetPinFragment(ostPinAcceptInterface);
+    }
+
+    @Override
+    public void invalidPin(long workflowId, OstWorkflowContext ostWorkflowContext, String userId, OstPinAcceptInterface ostPinAcceptInterface) {
+        showProgress(false);
+        showGetPinFragment(ostPinAcceptInterface);
+
+        Dialog dialog = DialogFactory.createSimpleOkErrorDialog(DashboardActivity.this,
+                "Incorrect PIN",
+                "Please enter your valid PIN to authorize");
+        dialog.setCancelable(false);
+        dialog.show();
+    }
 
     private void showGetPinFragment(OstPinAcceptInterface ostPinAcceptInterface) {
         WorkFlowPinFragment fragment = WorkFlowPinFragment.newInstance("Get Pin", getResources().getString(R.string.pin_sub_heading_get_pin));
@@ -318,13 +298,35 @@ public class DashboardActivity extends BaseActivity implements
     }
 
     @Override
-    public void popTopFragment() {
-        FragmentUtils.goBack(this);
+    public void pinValidated(long workflowId, OstWorkflowContext ostWorkflowContext, String userId) {
+
     }
 
     @Override
-    public void invalidPin(long workflowId, OstWorkflowContext ostWorkflowContext, String userId, OstPinAcceptInterface ostPinAcceptInterface) {
+    public void verifyData(long workflowId, OstWorkflowContext ostWorkflowContext, OstContextEntity ostContextEntity, OstVerifyDataInterface ostVerifyDataInterface) {
+        showProgress(false);
+        JSONObject jsonObject;
+        String dataToVerify = null;
+        WorkFlowVerifyDataFragment fragment = null;
+        if (OstSdk.DEVICE.equalsIgnoreCase(ostContextEntity.getEntityType())) {
+            fragment = VerifyDeviceDataFragment.newInstance();
+            OstDevice ostDevice = ((OstDevice) ostContextEntity.getEntity());
+            fragment.setDataToVerify(ostDevice);
+        } else {
+            fragment = VerifyTransactionDataFragment.newInstance();
+            jsonObject = (JSONObject) ostContextEntity.getEntity();
+            fragment.setDataToVerify(jsonObject);
+        }
 
+        fragment.setVerifyDataCallback(ostVerifyDataInterface);
+        FragmentUtils.addFragment(R.id.layout_container,
+                fragment,
+                this);
+    }
+
+    @Override
+    public void popTopFragment() {
+        FragmentUtils.goBack(this);
     }
 
     @Override
@@ -357,105 +359,29 @@ public class DashboardActivity extends BaseActivity implements
 
     @Override
     public void onDeviceSelectToRevoke(Device device) {
-//        WorkFlowListener revokeDeviceWorkflowListener = SdkInteract.getInstance().newWorkFlowListener();
-//
-//        OstSdk.revokeDevice(
-//                AppProvider.get().getCurrentUser().getOstUserId(),
-//                device.getDeviceAddress(),
-//                revokeDeviceWorkflowListener
-//        );
+        WorkFlowListener revokeDeviceWorkflowListener = SdkInteract.getInstance().newWorkFlowListener();
+
+        OstSdk.revokeDevice(
+                AppProvider.get().getCurrentUser().getOstUserId(),
+                device.getDeviceAddress(),
+                revokeDeviceWorkflowListener
+        );
     }
 
     @Override
     public void onDeviceSelectedForRecovery(Device device) {
-        String deviceAddress = null;
-        if (null != device) deviceAddress = device.getDeviceAddress();
-
-        String workflowId = OstWalletUI.initiateDeviceRecovery(this,
-                AppProvider.get().getCurrentUser().getOstUserId(),
-                deviceAddress,
-                AppProvider.get().getUserPassphraseCallback());
-        OstWalletUI.subscribe(workflowId, this);
-    }
-
-    @Override
-    public void initiateDeviceRecovery() {
-        onDeviceSelectedForRecovery(null);
-    }
-
-    @Override
-    public void resetPin() {
-        String workflowId = OstWalletUI.resetPin(this,
-                AppProvider.get().getCurrentUser().getOstUserId(),
-                AppProvider.get().getUserPassphraseCallback());
-        OstWalletUI.subscribe(workflowId, this);
-    }
-
-    @Override
-    public void updateBiometricPreference(boolean enable) {
-        String userId = AppProvider.get().getCurrentUser().getOstUserId();
-
-        String workflowId = OstWalletUI.updateBiometricPreference(this,
-                userId,
-                enable,
-                AppProvider.get().getUserPassphraseCallback());
-        OstWalletUI.subscribe(workflowId, this);
-    }
-
-    @Override
-    public void viewMnemonics() {
-        String workflowId = OstWalletUI.getDeviceMnemonics(this,
-                AppProvider.get().getCurrentUser().getOstUserId(),
-                AppProvider.get().getUserPassphraseCallback());
-        OstWalletUI.subscribe(workflowId, this);
-    }
-
-    @Override
-    public void revokeDevice() {
-        String workflowId = OstWalletUI.revokeDevice(this,
-                AppProvider.get().getCurrentUser().getOstUserId(),
-                null,
-                AppProvider.get().getUserPassphraseCallback());
-        OstWalletUI.subscribe(workflowId, this);
-    }
-
-    @Override
-    public void authorizeDeviceWithMnemonics() {
-        String workflowId = OstWalletUI.authorizeCurrentDeviceWithMnemonics(this,
-                AppProvider.get().getCurrentUser().getOstUserId(),
-                AppProvider.get().getUserPassphraseCallback());
-        OstWalletUI.subscribe(workflowId, this);
-    }
-
-    @Override
-    public void showAddDeviceQR() {
-        String workflowId = OstWalletUI.getAddDeviceQRCode(this,
-                AppProvider.get().getCurrentUser().getOstUserId());
-        OstWalletUI.subscribe(workflowId, this);
-    }
-
-    @Override
-    public void authorizeDeviceViaQR() {
-        String workflowId = OstWalletUI.scanQRCodeToAuthorizeDevice(this,
-                AppProvider.get().getCurrentUser().getOstUserId(),
-                AppProvider.get().getUserPassphraseCallback());
-        OstWalletUI.subscribe(workflowId, this);
-    }
-
-    @Override
-    public void executeTransactionViaQR() {
-        String workflowId = OstWalletUI.scanQRCodeToExecuteTransaction(this,
-                AppProvider.get().getCurrentUser().getOstUserId(),
-                AppProvider.get().getUserPassphraseCallback());
-        OstWalletUI.subscribe(workflowId, this);
+        Fragment fragment = InitiateRecoveryFragment.newInstance(device.getDeviceAddress());
+        FragmentUtils.addFragment(R.id.layout_container,
+                fragment,
+                this);
     }
 
     @Override
     public void onDeviceSelectedToAbortRecovery(Device device) {
-        String workflowId = OstWalletUI.abortDeviceRecovery(this,
-                AppProvider.get().getCurrentUser().getOstUserId(),
-                AppProvider.get().getUserPassphraseCallback());
-        OstWalletUI.subscribe(workflowId, this);
+        Fragment fragment = AbortRecoveryFragment.newInstance(device.getDeviceAddress());
+        FragmentUtils.addFragment(R.id.layout_container,
+                fragment,
+                this);
     }
 
     @Override
@@ -577,91 +503,5 @@ public class DashboardActivity extends BaseActivity implements
                 }
             }
         });
-    }
-
-    @Override
-    public void createSession(String spendingLimit, long expiryTime) {
-        String workflowId = OstWalletUI.createSession(this,
-                AppProvider.get().getCurrentUser().getOstUserId(),
-                expiryTime,
-                spendingLimit,
-                AppProvider.get().getUserPassphraseCallback());
-        OstWalletUI.subscribe(workflowId, this);
-    }
-
-    @Override
-    public void requestAcknowledged(OstWorkflowContext ostWorkflowContext, OstContextEntity ostContextEntity) {
-        if (OstWorkflowContext.WORKFLOW_TYPE.ADD_SESSION.equals(ostWorkflowContext.getWorkflowType())) {
-            showToastMessage("Session authorization request received.", true);
-        } else if (OstWorkflowContext.WORKFLOW_TYPE.REVOKE_DEVICE.equals(ostWorkflowContext.getWorkflowType())) {
-            showToastMessage("Device Revoke request received.", true);
-        } else if (OstWorkflowContext.WORKFLOW_TYPE.AUTHORIZE_DEVICE_WITH_MNEMONICS.equals(ostWorkflowContext.getWorkflowType())) {
-            showToastMessage("Authorize device request received", true);
-        }
-    }
-
-    private void logSdkError(OstWorkflowContext ostWorkflowContext, OstError ostError) {
-        StringBuilder errorStringBuilder = new StringBuilder();
-
-        String errorString = String.format("Error: %s " +
-                        "\nwith error code: %s" +
-                        "\ninternal error code: %s",
-                ostError.getMessage(),
-                ostError.getErrorCode(),
-                ostError.getInternalErrorCode()
-        );
-        errorStringBuilder.append(errorString);
-
-        if (ostError.isApiError()) {
-            OstApiError ostApiError = ((OstApiError) ostError);
-            String apiErrorCodeMsg = String.format(
-                    "\n%s: %s",
-                    ostApiError.getErrCode(),
-                    ostApiError.getErrMsg());
-
-            errorStringBuilder.append(apiErrorCodeMsg);
-
-            errorStringBuilder.append(
-                    String.format("\napi_internal_id: %s", ostApiError.getApiInternalId())
-            );
-
-            List<OstApiError.ApiErrorData> apiErrorDataList = ostApiError.getErrorData();
-            for (OstApiError.ApiErrorData apiErrorData : apiErrorDataList) {
-                String errorData = String.format(
-                        "\n%s: %s",
-                        apiErrorData.getParameter(),
-                        apiErrorData.getMsg());
-
-                errorStringBuilder.append(errorData);
-            }
-        }
-        String stringFormattedError = errorStringBuilder.toString();
-        Log.e(LOG_TAG, stringFormattedError);
-
-    }
-
-    private void deviceUnauthorized(OstError ostError) {
-        String title = "Device not registered";
-        String message = "Please login again to register your device.";
-        if (ostError.isApiError()) {
-            OstApiError apiError = (OstApiError) ostError;
-            if (apiError.isApiSignerUnauthorized()) {
-                title = "Device Revoked";
-            }
-        }
-        Log.e(LOG_TAG, title);
-        showLogoutDialog(title, message);
-    }
-
-    public void showLogoutDialog(String title, String message) {
-        AlertDialog.Builder builder = new AlertDialog.Builder(AppProvider.get().getCurrentActivity());
-        builder.setTitle(title);
-        builder.setMessage(message);
-
-        builder.setPositiveButton("Logout", new DialogInterface.OnClickListener() {
-            public void onClick(DialogInterface dialog, int id) {
-                AppProvider.get().relaunchApp();
-            }});
-        builder.create().show();
     }
 }